webapi/webapi-webspeech-w3c-tests/webspeech/speechrecognition_continuous-manual.html

<!DOCTYPE html>
<!--
Copyright (c) 2013 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Shentu. Jiazhen <jiazhenx.shentu@intel.com>

-->

<meta charset="utf-8">
<title>Web Speech API Test: SpeechRecognition continuous attribute</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="http://dvcs.w3.org/hg/speech-api/raw-file/tip/speechapi.html">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<p>Speak 3 words in 1 second</p>
<div id="log"></div>
<script>

var t = async_test("Test if the user agent must return zero or more final results representing multiple consecutive recognitions in response to starting recognition when the continuous attribute is set to true", { timeout: 5000});

t.step(function () {
  var speechReco = new webkitSpeechRecognition();
  speechReco.continuous = true;
  var finalNum = 0;
  speechReco.onresult = function (evt) {
    for (var i = 0; i < evt.results.length; ++i) {
      for (var m = 0; m < evt.results[i].length; ++m) {
        if (event.results[i][m].final) {
          finalNum++;
        }
      }
    }
  };
  speechReco.onspeechend = t.step_func(function () {
    assert_greater_than_equal(finalNum, 0);
    t.done();
  });
  speechReco.onerror = t.step_func(function (evt) {
    assert_unreached("SpeechRecognition get error " + evt.error);
    t.done();
  });
  speechReco.start();
  setTimeout(function () {
    speechReco.stop();
  }, 1000);
});

</script>
